Update: Gave more examples to what `call out boxes` is
I've added three more examples to what it is, though maybe I'm wrong myself. 

I've done this for a while and I've never heard what a call-out box is, seems more like a marketing terminology or something and I feel like more examples of what that is, could be beneficial. 

And while we do have `<dialog>` I'm not even sure if that is meant to be used for modals based on what I read about the `<dialog>` tag. 

I think being more specific would help a lot instead of generalizing these things. 

Perhaps adding even more examples where, and in dialog say that it woudln't be suitable for a 'modal', if its not, maybe someone can correct me here if I'm wrong.