Refer to timing properties by property name
`KeyframeEffect()` doesn't actually define "iteration duration" or
"iteration count". It also isn't a particularly good source for timing
property documentation, both because it uses the timing properties
somewhat differently (as optional parameters) and because in the future
it might be one of several effect constructors.
`AnimationEffect.getTiming()` would be a better source.