Be more precise about activeDuration
`activeDuration` is not actually the duration the effect will "run".
`delay` and `endDelay` may be negative and thereby overlap the
active duration and prevent the effect from running during the overlap.
Also, `fillMode` may cause the animation's effects to be applied
outside the `activeDuration`.

`activeDuration` is zero when either `duration` or `iterations` is
zero.