Update index.md
Changed the dimensions of the example to make the displayed area proportional to the shape of the SVG viewport. The example demonstrates shifting and resizing in an internal svg element, and it is easier to see how these work when there is no empty space included in the display.  Unfortunately, the preview does not ensure that the change I made will have the desired effect since the example is not shown and I did not see a way to test it.