refactor(prettier): Reverse .prettierignore to use as breakdown list
This reverses the Prettier ignore, so that instead of ignoring all Markdown and making exceptions for folders we have performed a pass on, we only ignore the folders that are YET to be formatted.  This allows us to use the `.prettierignore` as a breakdown list of what still needs to be performed.