Reword/restyle Glossary Definition for Speed Index
Hello, I am studying Speed Index right now and I came across this Glossary Definition. I thought that it could be slightly improved, with a little context about why this metric exists, and if it is appropriate to link to the two major projects using the metric in their audits.

I also think visually the image is large for the entry, so I opted to put all the text above the image with a description to further help.

Hope this proposed change resonates. I believe the information is correct at least, style wise maybe an improvement.