improve ambiguous sentences
I think the sentence `Classes in JS are built on prototypes but also have some syntax and semantics that are unique to classes.` doesn't make sense. It explains the difference `Classes in JS` and `classes`. I don't know exactly the intention of the original author, but when I inferred the link below([Using classes](/en-US/docs/Web/JavaScript/Guide/Using_Classes)), it seems to be correct to change it like this.