Fix link to next content page (#24654)
Update index.md

Currently, this page states, in the summary, that the next topic of study is the box model. 

However, the 'next' button is linked to the Cascade_layers page, and topics in the sidebar on the left are listed in this order:
-Cascade, specificity, and inheritance
-Cascade layers
-The box model

This could be intentional, but it feels strange to have a link after the words "Next up" and a link on the "Next" button go to different pages.

These are also redundant links now, so maybe the best approach is to remove the link in the summary statement altogether.