Fix ellipse angles definition
As was found in https://github.com/whatwg/html/issues/8408 the ellipse's `startAngle` and `endAngle` are not really angles on the ellipse, but angles on the eccentric circle of the ellipse.   
These are known as "eccentric angle", though this term *might* be a bit obscure at first sight, so I wouldn't mind a better proposition if reviewers have one.  
The specs got fixed in https://github.com/whatwg/html/pull/8495 but as can be seen, the fix was quite verbose and I'm not sure the docs need to get that much details.  
So I hope this little fix is both enough to lead the ones that could get surprised by the actual behavior to the proper understanding, and not too obscure to lose in details the ones that wouldn't notice.