Update index.md
Removes incorrect information the `separator` parameters information. The incorrect information is "so omitting it or passing undefined causes split() to split by the string "undefined", which is rarely what you want.", because omitting the `separator` argument causes `split` to return a new array with the calling string as a singular element.