Add cone mosaic, and
This is some updates to the area relating to color perception, I tried very hard to keep the spirit of what and how it was written, simply trying to correct some errata. Also added an image of the cone mosaic to show the typical dispersement of cones, one for StandardVision and one for protanopia. Please let me know if you have any questions or if you feel it needs additional references or edits.