Revert local and session storage quotas to 5MiB
Based on a discussion with Andrew Sutherland:

our quota measurement and enforcement is on the string length of 5 * 1024 * 1024 code units across all keys and values, not on the actual disk usage. Because JS uses 16-bit code units, this ends up being capable of storing 10 MiB of data. But I think phrasing it "10 MiB" is going to lead to people making the wrong inference that they can put 10 * 1024 * 1024 code units worth of data in there."our quota measurement and enforcement is on the string length of 5 * 1024 * 1024 code units across all keys and values, not on the actual disk usage. Because JS uses 16-bit code units, this ends up being capable of storing 10 MiB of data. But I think phrasing it "10 MiB" is going to lead to people making the wrong inference that they can put 10 * 1024 * 1024 code units worth of data in there."