Reference for...of statement before usage in example (#24715)
* Reference for...of statement before usage in example

The 'Showing file(s) size' example includes a for...of statement to iterate through the file array. However, this has not been introduced previously in the page (instead a for loop with an iteration variable was used just above). To prevent potential confusion, this change mentions the for...of statement and gives a snippet of how to use it before it's referenced.

Alternatively, I'd suggest changing the 'Showing file(s) size' example to use a for loop with an iteration variable for consistency.

* Remove file iteration examples

Readers will be familiar with looping through arrays, so the `for (;;)` and `for...of` snippets are not required.