Added arcTo() details and demo
The demo that was added is a useful way to see how arcTo() works. However, the code is more related to the user interactions than arcTo(), so it was hidden using the methods described at https://developer.mozilla.org/en-US/docs/MDN/Writing_guidelines/Page_structures/Code_examples. So the HTML should be hidden, but it appears in the preview. Perhaps this is a bug in the previewer?