Revamp the storage quota and eviction documentation (#25019)
* Moving the storage page to the learn area

* New storage quota/eviction page content

* Removed TODO

* Update files/en-us/learn/javascript/client-side_web_apis/browser_storage_limits_and_eviction_criteria/index.md

Co-authored-by: dawei-wang <dawei-wang@users.noreply.github.com>

* Update files/en-us/learn/javascript/client-side_web_apis/browser_storage_limits_and_eviction_criteria/index.md

Co-authored-by: dawei-wang <dawei-wang@users.noreply.github.com>

* Update files/en-us/learn/javascript/client-side_web_apis/browser_storage_limits_and_eviction_criteria/index.md

Co-authored-by: dawei-wang <dawei-wang@users.noreply.github.com>

* Addressing Estelle's comments

* Updated the web storage quotas

* Addressed Ayu's comments

* Address Andrew's comments

* Moving page to Storage_API

* Fixed file path

* Fixed sidebar

* Re-worked the Storage_API page

* Fixed links

* Revert local and session storage quotas to 5MiB

Based on a discussion with Andrew Sutherland:

our quota measurement and enforcement is on the string length of 5 * 1024 * 1024 code units across all keys and values, not on the actual disk usage. Because JS uses 16-bit code units, this ends up being capable of storing 10 MiB of data. But I think phrasing it "10 MiB" is going to lead to people making the wrong inference that they can put 10 * 1024 * 1024 code units worth of data in there."our quota measurement and enforcement is on the string length of 5 * 1024 * 1024 code units across all keys and values, not on the actual disk usage. Because JS uses 16-bit code units, this ends up being capable of storing 10 MiB of data. But I think phrasing it "10 MiB" is going to lead to people making the wrong inference that they can put 10 * 1024 * 1024 code units worth of data in there."

* Addressed Will's comments

* Clarified total disk space

* Addressed Firefox's group limit comment

---------

Co-authored-by: dawei-wang <dawei-wang@users.noreply.github.com>