Update index.md
Hello again! I saw a typo in my suggestion 🙄. Fixed and worded a little differently! 

I've been having 100% success with the "try it yourself" stuff in the guide (starting from 0 in HTML, thru the end of the CSS section. This is the one place I was really scratching my head... It really felt like most of the css was not applying here. While I was writing you an issue on GH, I accidentally found why it wasn't working -- my window was too narrow. All I saw was the original column of text, but a little prettier.

Might be nice to have some sort of warning about this on the page. I'm sure you all can find better words, though! :)

Btw, this guide has been THE BOMB. Thank you so much for doing what you all do! Seriously!! The embedded videos are great too!