Add suggestion for IIFEs to avoid with statements
The documentation for the `with` statement wisely advises its avoidance, and does a good job of explaining why. Still, people who show up on this page may be seeking a `with`-like solution without its drawbacks. One such solution is already offered, and it's good: destructuring. This change-set introduces another, which is more appropriate to those who are producing an expression.