Updated index.md
On line 114, it mentions forms being included in the HTML part of this course.  I'm in the middle of going through the content and I didn't encounter this so I thought I skipped over a section by mistake. I didn't. 

The content for the web forms section is further down from the CSS section.  I suggest that this paragraph should refer to the web forms section instead of the HTML section. This will be less confusing for new learners.