Fix nginx CORS instructions
The "always" option was missing, resulting in the "Access-Control-Allow-Origin"
header not being added for many HTTP status codes.

For example, this would prevent a client to read the content of a 404 Not Found response.

Further, get rid of the 'origin-list` placeholder.
It suggests that a list of URLs can be given, which is wrong:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin

> Only a single origin can be specified.

Finally, switch example domain to example.com.