Specify webRequest.onAuthRequired prerequisites
The `webRequest.onAuthRequired` event is only fired when the server responds with 401/407 and the `WWW-Authenticate` header uses the `Basic` scheme. This second precondition is used by the browser to prompt the user for credentials. However, it wasn't clear in the docs that other authentication schemes are ignored by the webRequest API.