similar is identical (#26401)
A one word change that clarifies that the timeStamp argument passed to your requestAnimationFrame callback function is the same value that you would get if you called performance.now() at the same time as you called the callback.  If you check performance.now() within a callback function the value is only a few milliseconds later.\
I do this now because I needed a way to get a timeStamp arg value without running requestAnimationFrame, and performance.now() did the trick.  I needed it because I am allowing user-pausing and resuming of an animation, and I need to reset the start time based on the elapsed duration of the user pause.  It helped to be able do it outside of the rAF callback loop.