Update index.md
Line 63, should be style.color so that the color changes with the div that it reads the style off of- as described in the sample description.  The sample hosted @ https://mabian-ms.github.io/delegated-ink-trail.html needs to be updated as well.

With this change the dots following the mouse match the color  of the div for 50 ticks.  Works fine before, just doesnt achieve what it is stated to achieve.