Update index.md
In the example for the import of the public part of the rsa key there is a small error which prevents the proper base64 decoding. When the part of the pem string is fetched from between the header and the footer, the length of the string is one character too long, a '-' is still trailing. So I added '- 1', then the pem string can be decoded.