Add missing heading to static HTML
It's specified in the breakdown (line 127) and shows up in the screenshot (line 244), but is oddly missing from the "complete" version.

I looked to see if it was intentionally removed at some point, but that doesn't seem to be the case. I went back to commit 1ce41dbdd4e0888f453c29ae349759e34fcd47e9 which seems to be the original commit and the heading seems to just have been accidentally left out.