Remove dash in margin box to match W3 spec
In https://drafts.csswg.org/css-box-4/#margin-edge the area of the margin edges is described as the margin box, without a dash. margin-box is used elsewhere in https://drafts.csswg.org/css-box-4/#keywords as a keyword for some box edges. I think the use of it on this page is much more similar to the first example so probably shouldn't use a dash.