Remove invalid references to margin-box (#28157)
* Remove references to margin-box

margin-box was removed in the 2021 5 August 2021 Candidate Reccomendation, see https://drafts.fxtf.org/css-masking-1/#changes

* Remove margin-box from mask-clip

* Replace Geometry box with Coord box

Geometry box is from an old spec, and has since been replaced with Coord box. Additionally only some of the possible values were listed, now all of them are

* Remove dash in margin box to match W3 spec

In https://drafts.csswg.org/css-box-4/#margin-edge the area of the margin edges is described as the margin box, without a dash. margin-box is used elsewhere in https://drafts.csswg.org/css-box-4/#keywords as a keyword for some box edges. I think the use of it on this page is much more similar to the first example so probably shouldn't use a dash.

---------

Co-authored-by: Chris Mills <chrisdavidmills@gmail.com>