Reword note about pushState() state object size limit
Fixes https://github.com/mdn/content/issues/28516.

This change takes some text about a particular size limit in Firefox for
the pushState() state object, and rewords that text for better
maintainability — both because the docs currently state two very
different, conflicting values for that size limit, and also because
stating any explicit size limit is bad for maintainability, because the
limit might change (and invalidate what we state in the docs).