CycleTracker Tutorial Proofread: HTML and CSS (#28267)
* Remove unrendered line break

When rendered on the web, the single line break is taken out. My intent is to remove any potential confusion about the end result.

* Separate long clause for clarity

* Thru -> Through

* Separate appositive to avoid ambiguous meaning of comma

* Replace en dash with em dash.

Refs: https://www.scribbr.com/language-rules/dashes

* Add missing article

* Add missing 'n' in "accidetally"

* Use link on _first_ label instead of second

* Adjust parenthesis

"two paragraphs elements" doesn't make sense, so I figure it was intended to be "two paragraphs (<p> elements)"

* Separate what button does from what we do

* Fix typo in "cycles"

* Add tip about previewing unstyled HTML

* Make css explanation easier to read

* Include placeholder past-period section in "complete" file.

For some reason, it wasn't included before.

* Add a missing 'to'

* Add missing heading to static HTML

It's specified in the breakdown (line 127) and shows up in the screenshot (line 244), but is oddly missing from the "complete" version.

I looked to see if it was intentionally removed at some point, but that doesn't seem to be the case. I went back to commit 1ce41dbdd4e0888f453c29ae349759e34fcd47e9 which seems to be the original commit and the heading seems to just have been accidentally left out.

* Add missing 'a'

* Revert "Remove unrendered line break"

This reverts commit e1a31e4de9428debfffb64c57d7836fb89d3e6e6.

* Revert "Separate long clause for clarity"

This reverts commit 7fded77bc3ed26caba296d5cf8e9dc9fde81d3bc.

* Revert "Use link on _first_ label instead of second"

This reverts commit d1ae6c07a785efa94b3cc11965a19bb1ef9631fc.

* Revert "Include placeholder past-period section in "complete" file."

This reverts commit 49fab78cc47a91e871d68f369e8f662417871ad7.

* Unreverse "we" and "will"

* Update files/en-us/web/progressive_web_apps/tutorials/cycletracker/html_and_css/index.md

---------

Co-authored-by: Brian Thomas Smith <brian@smith.berlin>