restore link to outdated vanilla framework
back in #25383 the vanilla framework for xhr uploads was removed, i don't really have any context for that decision, but it left a hole in the documentation. 

this pr updates the existing link in the formdata docs to the old documentation. 

that framework _is useful_ in the context of this document. i don't know how mdn would ideally extract that framework, but this at least restores the context for that link.