Update index.md
It conveys the information that only multi core processors can help with multi threading which as far as I know is incorrect. The concept of multithreading was there even in single core processors. 

However, the advantage that multicore processors offer is that it can allow for a thread to be executed in in another core when any core of the multicore processor is free.