Update index.md, "Loading only the glyphs you need" (#28771)
Update index.md, correcting issues with "Loading only the glyphs you need"

The section with the heading "Loading only the glyphs you need" contains language that incorrectly suggests that `unicode-range` controls **what** is downloaded. However it only controls **if** a font is downloaded. This is my attempt to correct the language.