Fix Promise.prototype.then's parameter description
It seems like that the description of onRejected, which is the second
parameter of Promise.prototype.then, was copied from the
Promise.prototype.catch page. This likely happened due to the same
functionality of both onRejected parameters.

However the original author forget to replace 'catch' in the
description with 'then'. This patch fixes this minor issue.