Update index.md  indexEnd is not the end since we actually end before it
For long the slice has been a source of confusion among developers. Apart from it's similarity with the splice in terms of name , I feel the naming of the parameters has also contributed. Also I had to specify what happens when we have negative arguments because clearly what happens with single negative arguments is different from what happens generally with the slice. Also naming our second parameter indexEnd is confusing because that particular index is not a part of it , as long as we counting from 0.