Clarify the nesting spec the example shows
The "Without nesting selector" section is confusing. The code shown here aims to show the impact of omitting the nesting selector (&).  However, the section does not mention that the point being made applies to the old nesting spec--the two examples shown produce the same output with FireFox.  To get the results shown you need be using Chrome or Safari (as of 12 September).