Clarify the nesting spec the example shows (#29074)
* Clarify the nesting spec the example shows

The "Without nesting selector" section is confusing. The code shown here aims to show the impact of omitting the nesting selector (&).  However, the section does not mention that the point being made applies to the old nesting spec--the two examples shown produce the same output with FireFox.  To get the results shown you need be using Chrome or Safari (as of 12 September).

* fix suggested lint error

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* Update files/en-us/web/css/css_nesting/using_css_nesting/index.md

* Update files/en-us/web/css/css_nesting/using_css_nesting/index.md

---------

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>
Co-authored-by: Estelle Weyl <estelle@weyl.org>