Update index.md
Edited line 225 to show how the static portion of the web service plays a key role in presenting the initial user interface for a user to start interacting. I understand this captures the vital role of the static resource and makes it immediately clearer to the reader of its importance.