Make onOutOfBounds.add(... consistent
The example code in the body of the page for checking if the ball is out of bounds is different than the code in the "Compare your code" section. While they both work, the add(function() ... is clearer (IMHO) to new developers than the add(() => { ... syntax.

This commit changes line 27 from the () => syntax to function().