Move example oklch colours closer to gamut bounds
The differing absolute ranges of different coordinates in OKLCH can be confusing at first, and these examples are using a chroma value of 130, which is extremely far out of gamut. 0.5 is also out of gamut for this particular colour, but is within the rough 0–0.5 range in which one generally works, mentioned earlier on the page.