addEventListener: Update example function name to prevent minor confusion
The notes above this example teach that the second argument of the `addEventListener` function can accept an object that has a `handleEvent()` function. The example then shows what that might look like but names the function `eventHandler`.

This can cause a little bit of confusion so this PR just makes a minor rename to mitigate that.