HTMLAnchorElement "hash" property: the fragment *is* percent-encoded
The documentation was stating the contrary of the actual encoding state. The fragment *is* percent-encoded, actually.

As described in the compatibility table, from Firefox 29 to Firefox 40, the returned value was incorrectly percent-decoded (i.e. not percent-encoded).

Firefox 29 was released on April 29, 2014. Firefox 41 was released on September 22, 2015. Now we are 8 years later.