refX and refY set to the tip of the arrowhead
This change would only apply to arrowhead marker. Setting 'refX' and 'refY' to the tip of the arrow head, that is the vertex (10,5) of the triangle, gives a more accurate representation of the arrow because it doesn't increase its length when adding the arrowhead. This is convenient for vectors representation.