Extend the getTargetRanges method documentation
Introduce a new table mirroring an existing table in the W3C spec
that explains the expected return values for the getTargetRanges
method. Notably, this table highlights that for editing hosts that
are not contenteditable, the expected return value for this method
is an empty array (which is not currently clear anywhere except the
W3C docs).