docs(hgroup): adds related note from a different page
I was reading through this page:
https://developer.mozilla.org/en-US/docs/Web/HTML/Content_categories

when I came across this note:

Note: The <hgroup> element is not recommended as it does not work properly with assistive technologies. It was removed from the W3C HTML specification prior to HTML 5 being finalized, but is still part of the WHATWG specification and is at least partially supported by most browsers.

I believe it's important to include this note in <hgroup>'s page because as it stands right now it might mislead developers into using it, thinking that it has no drawbacks.

Someone might use <hgroup> instead of a <div> due to preference, so they can use it in a css selector or to create a reusable pre-styled container and unknowingly make their website less acessible.