Added arcTo() details and demo (#25348)
* Added arcTo() details and demo

The demo that was added is a useful way to see how arcTo() works. However, the code is more related to the user interactions than arcTo(), so it was hidden using the methods described at https://developer.mozilla.org/en-US/docs/MDN/Writing_guidelines/Page_structures/Code_examples. So the HTML should be hidden, but it appears in the preview. Perhaps this is a bug in the previewer?

* chore(api): Run example through formatter

* chore(api): Edits for arcTo() 2D canvas method examples

* chore(api): Incorporate hints from Hamish about Exceptions

---------

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>
Co-authored-by: Queen Vinyl Da.i'gyu-Kazotetsu <vinyldarkscratch@gmail.com>
Co-authored-by: Brian Thomas Smith <brian@smith.berlin>