docs(css): Add details about number of properties and values in CSS syntax (#31094)
replace "100 or more" with "hundreds". Learning 100-200 properties seemed possible to me, but on following the link there were actually hundreds of properties, too many for me to learn.

replace "almost infinite" with "practically endless". In a context where precise meaning is important, "almost infinite" is distracting, since it makes no sense if it is interpreted literally.