Update index.md
In the file I have corrected a typo at the "hiding a column by changing style properties" instead of the "col === 0" I put "i === 0" because the variable "col" is not defined in the script. Thus, using col the code was not working. Also,, at the "Getting a text node from the table" there was a type at part b) (I am not sure of the change though.