Show that objects usually inherit the `constructor` property
The paragraph above my change says this in its first sentence:

> Note that `constructor` usually comes from the constructor's [`prototype`](/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/prototype) property.

But the example that follows doesn't *show* that. Therefore, I added `o.hasOwnProperty("constructor"); // false` to the example. It might make more sense to put it as the second line instead of where I placed it? 

I also wrote this first but removed it later: `o.constructor === proto(o).constructor // true`: prototypical inheritance can be tricky to people like me, and I felt that this line might not document its intention as well as `o.hasOwnProperty`. Let me know if that's a better line or if both should be in the example.