Add back decimal in color, and add to rgb/hsl
I removed initially because it looked out of place, but actually it was probably to demonstrate having a decimal value (since it wasn't always supported in rgb). I've added it to rgb as well as rgba, and added an equivalent to hsl (in addition to hsla) to make it obvious that both versions work with decimals.