color syntax - make alpha values consistent
I don't think there was any reason to have different alpha values for these two rgba examples. The different syntaxes are to show how they're used, so changing the alpha value for just rgba seemed to imply that it somehow uses alpha differently from rgb.

This also brings it in line with the hsl/a examples which are consistent (also with 60% opacity)