Remove prefixed :read-only selectors (#31718)
If a browser supports either `:read-only` or `:-moz-read-only`, but not both, then the entire ruleset was discarded, as the selector list is unforgiving.

Actually, Firefox is the only browser that supports both. All other browsers were discarding the ruleset.

Possible fixes are:
* Use two separate rulesets (one with the prefixed selectors, another with the unprefixed selectors).
* Enclose the selectors in a `:is()` pseudo-class, as its selector list is forgiving.

Though, Firefox supports the unprefixed selector since version 78, released on June 2020. So we may just drop the prefixed selectors.