Update remaining web area to use latest rgb() and hsl() syntax (#31497)
* Update rgb/hsl syntax in web\accessibility

* Update rgb & hsl to use new syntax

Including percentage instead of number for alpha value

* Replace instances of rgba() with rgb() in web\api

Includes rewording of canvas_api\tutorial\applying)styles to use rgb instead of rgba and include a percentage value for alpha. Rename referenced canvas_rgba.png → canvas_rgb.png.

There weren't any instances of hsla()

* Update rgb/hsl syntax in web\guide

* Update rgb/hsl syntax in web\html

Remove reference to rgba in element\input\color

* Update rgb/hsl syntax in web\mathml

I've left mathcolor as it lists all the possible ways of using color, including rgb and rgba, new and old syntax.

* Update rgb/hsl syntax in web\mediia

* Update rgb/hsl syntax in web\progressive_web_apps

* Update rgb/hsl syntax in web\svg

* Update miscellaneous rgb/hsl syntax in web

Including fixing a typo in css\color

* Apply linter/prettier suggestions

* Fix some incorrect syntax found in review

* Update files/en-us/web/html/element/input/color/index.md

Co-authored-by: shubham oulkar <91728992+ShubhamOulkar@users.noreply.github.com>

* Update files/en-us/web/api/mutationrecord/oldvalue/index.md

Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>

* Add 'with alpha transparency' to rgb() example

This is because this section is about using rgb() to specify alpha transparency instead of the globalAlpha property, so the focus should be on changing the transparency

* Update files/en-us/web/api/canvas_api/tutorial/applying_styles_and_colors/index.md

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

* Update files/en-us/web/api/canvas_api/tutorial/applying_styles_and_colors/index.md

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

* Update files/en-us/web/api/canvas_api/tutorial/basic_animations/index.md

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

* Fix incorrect rgb syntax in html area

* Rename rgba to rgbColor

* Move related code onto the same line

* Move related code onto the same line

* Update files/en-us/web/api/canvas_api/tutorial/applying_styles_and_colors/index.md

Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>

* Add back comma in using_css_gradients

* Update comments in color

* Add back decimal in color, and add to rgb/hsl

I removed initially because it looked out of place, but actually it was probably to demonstrate having a decimal value (since it wasn't always supported in rgb). I've added it to rgb as well as rgba, and added an equivalent to hsl (in addition to hsla) to make it obvious that both versions work with decimals.

* color syntax - make alpha values consistent

I don't think there was any reason to have different alpha values for these two rgba examples. The different syntaxes are to show how they're used, so changing the alpha value for just rgba seemed to imply that it somehow uses alpha differently from rgb.

This also brings it in line with the hsl/a examples which are consistent (also with 60% opacity)

* Remove legacy rgba/hsla syntax from mathml\mathcolor

I decided to make the decimal example .6 / .2 rather than .0 as I think that's better at showing using them, and is also consistent with https://developer.mozilla.org/en-US/docs/Web/CSS/color

* Commit prettier suggestions

As strange as it appears to be

---------

Co-authored-by: shubham oulkar <91728992+ShubhamOulkar@users.noreply.github.com>
Co-authored-by: Estelle Weyl <estelle@openwebdocs.org>
Co-authored-by: Dipika Bhattacharya <dipika@foss-community.org>