Update NuxtJS Link to Current Version
This PR addresses the issue of the outdated link to **NuxtJS**, https://nuxtjs.org, which currently redirects to https://v2.nuxt.com. 
This older link points to a version of NuxtJS (v2) that is nearing its end-of-life. The link has been updated to https://nuxt.com/, which leads to the latest version, NuxtJS v3.