Update index.md
you can also change this to something else like "on water" This is because if this is not done, iceCreamVanOutside will always be true. The message "You should leave the house quickly." is always displayed. For better understanding, it is best to ascribe a condition so that it can be compared to the string as this can be misleading to newbies