Fixed examples in linear easing function description (#32001)
* Fixed examples in linear easing function

The examples had the animation progress and duration reversed. Since which of these is which can get confusing, a specific case of 100 pixels and 100 seconds was used for additional clarification.

* clarified meaning of animation time

animation time -> animation duration time to explicitly distinguish from animation-delay and also to help make the connection between duration and time when duration is mentioned elsewhere.