Update passing arguments section
The paragraph below the example distinguishes the terms parameter and argument. The following paragraph is about how arguments are passed to functions. One would expect that there would be another comparison between these terms but this is not the case. In my opinion, it is more clear to say "Arguments in JavaScript are always".