Error in specificity weight calculation table
The table that demonstrates specificity weight calculation for some selectors contains an error. For the last selector, i.e. `button:not(#mainBtn, .cta)`, the total specificity is `1-1-1`. However, on the site, it's mentioned as `1-0-1`. This is incorrect. `button` accounts for the rightmost `1`, `#mainBtn` results in leftmost `1` and the middle `1` is due to `.cta`. 

Also, a minor correction, `button:not(#mainBtn, .cta`) has been replaced with `button:not(#mainBtn, .cta)`.