Update index.md
I was thinking of something like "to set the rate of increment and decrement." The word step is the best word to use instead of rate but explaining a word with that same word does not do much. Rate is good also, but people might get confused with its semantics.