Rewrite `finally()`'s "Return value" section
This was the original wording:

> Returns an equivalent Promise. If the handler throws an error or returns a rejected promise, the promise returned by finally() will be rejected with that value instead. Otherwise, the return value of the handler does not affect the state of the original promise.

I would've liked to have written this as pure prose, but I couldn't find a way to refer to the `promise` in `promise.finally(onFinally)` with words alone. I may lack the terminology needed to articulate it. Some ideas I experimented with were, "The promise that called `finally()` and "The previous link of the promise chain." Both felt too abstract and/or convoluted to me, so I worried it would confuse the reader. 

Fixes #32736