Decrease 'scariness' of Writing WebSocket Servers intro
The previous version of this introduction indicates that many people
find writing a custom web server "scary". Tweak the text lightly so it
acknowledges that it might seem *overwhelming*, but does not indicate to
readers that it could (still less imply that it *should*) seem scary,
and emphasize that it can be straightforward.