Add note about `ProgressEvent.loadel` reporting different values depending on the browser
In Firefox, `loaded` is the size of the compressed data; in Chrome, it's the size of the decompress data. I would have loved to find sources for this, but the only thing I could find is that the RFC doesn't specify which value it should be! Browsers implement it differently.