Deprecate OffscreenCanvas2D's commit() method
This method has been removed in https://github.com/whatwg/html/pull/9979

The placeholder <canvas>'s bitmap is now updated during the "update the rendering" steps of the Worker or Window's event-loop that owns the OffscreenCanvas object. Not sure how to formulate that clearly.

This method has been (partially) implemented in 2 browsers before it got removed, but at least Safari is in the process of removing it (https://github.com/WebKit/WebKit/pull/27196), and there may be another method with the same name coming later. So it's not clear if simply removing the method's page like that is fine or if it would be better to keep it with a deprecation notice instead?

Compat table would also need to be updated, but I'll wait for a position on the above question.

Also, while I was on that page I fixed an approximative wording, and the position of the roundRect() method.