Update web/javascript/guide/grammar_and_types
Consider variables declared with `var` keyword. During the creation phase they are stored in the variable object of their execution context with the default initialization value of undefined. So if we use them in their scope before their declaration, they won't throw the error `cannot access before initialization` and they are initialized with the default value of `undefined`. But they are not yet assigned a value.